fix: load bearer.yml file automatically
